refactor(date-picker): Interaction behavior in readonly mode
